requirements.txt
gunicorn>=22.0
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.6
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.4
python-dotenv>=1.0
Pillow>=10.0
